long

+++ src/Main.java
@@ -2,7 +2,7 @@ public class Main {
 
     public static void main(String[] args) {
 
-        SalesManager convert = new SalesManager(new int[]{10});
+        SalesManager convert = new SalesManager(new long[]{10});
 
         convert.max();
         System.out.println(convert.max());
